Move bulk version requests to public API package
Both of these classes are part of the public API and so should be in the public
API package.
