+++ vidarr-pluginapi/src/main/java/ca/on/oicr/gsi/vidarr/api/BulkVersionRequest.java
@@ -1,4 +1,4 @@
-package ca.on.oicr.gsi.vidarr.server.dto;
+package ca.on.oicr.gsi.vidarr.api;
 
 import java.util.List;
