+++ vidarr-pluginapi/src/main/java/ca/on/oicr/gsi/vidarr/api/BulkVersionUpdate.java
@@ -1,4 +1,4 @@
-package ca.on.oicr.gsi.vidarr.server.dto;
+package ca.on.oicr.gsi.vidarr.api;
 
 public final class BulkVersionUpdate {
