+++ vidarr-server/src/main/java/ca/on/oicr/gsi/vidarr/server/Main.java
@@ -43,6 +43,7 @@ import ca.on.oicr.gsi.vidarr.api.AddWorkflowRequest;
 import ca.on.oicr.gsi.vidarr.api.AddWorkflowVersionRequest;
 import ca.on.oicr.gsi.vidarr.api.AnalysisOutputType;
 import ca.on.oicr.gsi.vidarr.api.AnalysisProvenanceRequest;
+import ca.on.oicr.gsi.vidarr.api.BulkVersionRequest;
 import ca.on.oicr.gsi.vidarr.api.ExternalKey;
 import ca.on.oicr.gsi.vidarr.api.ExternalMultiVersionKey;
 import ca.on.oicr.gsi.vidarr.api.InFlightCountsByWorkflow;
@@ -67,7 +68,6 @@ import ca.on.oicr.gsi.vidarr.core.OperationStatus;
 import ca.on.oicr.gsi.vidarr.core.Phase;
 import ca.on.oicr.gsi.vidarr.core.Target;
 import ca.on.oicr.gsi.vidarr.server.DatabaseBackedProcessor.DeleteResultHandler;
-import ca.on.oicr.gsi.vidarr.server.dto.BulkVersionRequest;
 import ca.on.oicr.gsi.vidarr.server.dto.ServerConfiguration;
 import ca.on.oicr.gsi.vidarr.server.remote.RemoteInputProvisionerProvider;
 import ca.on.oicr.gsi.vidarr.server.remote.RemoteOutputProvisionerProvider;
